+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* Marked as deprecated `ydb.MustParamsFromMap`
+* Removed experimental helper `ydb.MustParamsFromMap`
 * Changed result of experimental helper `ydb.ParamsFromMap` from tuple <`params.Parameters`, `error`> to `params.Parameters` only 
 
 ## v3.93.1