+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Removed experimental helper `ydb.MustParamsFromMap`
+* Changed result of experimental helper `ydb.ParamsFromMap` from tuple <`params.Parameters`, `error`> to `params.Parameters` only 
+
 ## v3.93.1
 * Published `query.ExecuteOption` as alias to `internal/query/options.Execute`