+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.93.2
 * Removed experimental helper `ydb.MustParamsFromMap`
 * Changed result of experimental helper `ydb.ParamsFromMap` from tuple <`params.Parameters`, `error`> to `params.Parameters` only